requirements.txt
numpy
pillow
timm
torch
